+++ examples/BooleanSwitching/Phase/BooleanSwitchingMaps.h
@@ -9,16 +9,16 @@
 #include "boost/foreach.hpp"
 #include <utility>
 
-std::vector < std::pair < int, int > >
-BooleanSwitchingMaps ( std::vector<int> const& closestface );
+std::vector < std::pair < int64_t, int64_t > >
+BooleanSwitchingMaps ( std::vector<int64_t> const& closestface );
 
-inline std::vector < std::pair < int, int > >
-BooleanSwitchingMaps ( std::vector<int> const& closestface ) {
-  std::vector < std::pair < int, int > > result;
-  std::vector < int > cold; 
-  std::vector < int > hot;
-  for ( int d = 0; d < closestface . size (); ++ d ) {
-    int axis = d + 1;
+inline std::vector < std::pair < int64_t, int64_t > >
+BooleanSwitchingMaps ( std::vector<int64_t> const& closestface ) {
+  std::vector < std::pair < int64_t, int64_t > > result;
+  std::vector < int64_t > cold; 
+  std::vector < int64_t > hot;
+  for ( int64_t d = 0; d < closestface . size (); ++ d ) {
+    int64_t axis = d + 1;
     switch ( closestface [ d ] ) {
     case 0:
       hot . push_back ( - axis );
@@ -38,8 +38,8 @@ BooleanSwitchingMaps ( std::vector<int> const& closestface ) {
     cold . push_back ( 0 );
     hot . push_back ( 0 );
   }
-  BOOST_FOREACH ( int c, cold ) {
-    BOOST_FOREACH ( int h, hot ) {
+  BOOST_FOREACH ( int64_t c, cold ) {
+    BOOST_FOREACH ( int64_t h, hot ) {
       result . push_back ( std::make_pair ( c, h ) );
     }
   }
